Move CameraController to it's own folder and update to work with StateMachine
